requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
line-bot-sdk>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0
pydub>=0.25.0

# LangChain & RAG
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.32
langchainhub>=0.1.0
qdrant-client>=1.7.0
ollama>=0.1.8

# 非同步 HTTP 與檔案處理
httpx>=0.24.0
aiofiles>=23.1.0

# Stripe 付費功能（如需商業化）
stripe>=8.0.0

# 開發/測試工具
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
